+++ model/src/main/java/com/indracompany/treinamento/model/dto/TransferenciaBancariaDTO.java
@@ -1,19 +1,22 @@
 package com.indracompany.treinamento.model.dto;
 
 import java.io.Serializable;
-
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
-public class TransferenciaBancariaDTO implements Serializable{
+public class TransferenciaBancariaDTO implements Serializable {
 
-	
-	private static final long serialVersionUID = 7198071248368049141L;
-	
-	private String agenciaOrigem;
-	private String numeroContaOrigem;
-	private String agenciaDestino;
-	private String numeroContaDestino;
-	private double valor;
+  private static final long serialVersionUID = 7198071248368049141L;
+
+  private String agenciaOrigem;
+
+  private String numeroContaOrigem;
+
+  private String agenciaDestino;
+
+  private String numeroContaDestino;
+
+  private BigDecimal valor;
 
 }
